convert to graphics coordinates
The MMSP grid data is stored row-major, so iterating over nodes(grid)
rasters over x first, y next, etc. So, a simple grid storing
 data -> abcdefghi  is read out as abc
                                   def
                                   ghi.

A column-major language, such as Fortran, would interpret this adg
                                                               beh
                                                               cfi.

To translate, simply swap the order of the loops: x last, y second-to-last,
etc. More succinctly, the column-major representation is the transpose of
the original data. VTK is column-major, so mmsp2vti and mmsp2pvd are
correctly implemented.

In addition to a column-major data format, graphics use a different
coordinate system than MMSP, so the same data array (abcdefghi)
interpreted by an image viewer produces:

   MMSP       PNG                        VTK
   y          +--->x = x     =     x     x
   ^ghi       |cfi     ^ihg     ghi^     ^ghi
   |def       |beh     |fed     def|     |def
   |abc       |adg     |cba     abc|     |abc
   +--->x     y        +--->y y<---+     +--->y

This is readily interpreted as amirroring about the y-axis of the
column-major representation, and implemented by copying the mmsp2vti
sequence but decrementing instead of incrementing the y-axis coordinate.

At least, that's the transformation that produced a clear picture :D